Add command to set desktop background in macos script
The original commands worked by replacing the symbolic link to the system's default background image. But because the system partition is not user writable any longer in macOS and because the default image is now in the HEIC format, use a command to replace the setting in the SQLite database instead.

Use the included "Stone" solid color image.